Finish the evaluation split and drop the duplicated protocol
The evaluation modules were carved out of a 2500-line file in 981656e,
which left four seams unfinished. The harness imported _chart_traces and
friends out of the HTML renderer to produce data that goes in the results
JSON, so data production depended on presentation; those move to
eval_metrics beside _compute_metrics, and _Sample -- produced by the
harness, read by both -- moves to eval_spec. render_html_report took six
parameters its only caller always filled from eval_compare, so the CLI
was assembling the report; it now takes (base, head). The 30-name
re-export shim in evaluation.py served one test file, which now imports
from the real modules, and the skill sentence claiming CI needs those
re-exports is corrected here too.

simulator/protocol.py said it existed so the simulator imports nothing
from the main package, but three sibling modules import astrameter.ct002
today. It was ct002/protocol.py modulo annotations, so it is gone; the
one call site passes the payload straight to sendto, which takes any
bytes-like.

steering_common.py holds _f32, _share_split and the three spike
constants, which the firmware and venus modules had verbatim copies of.
Their gate skeletons stay separate -- the one-shot latch versus
indefinite suppression is a documented open question.

Also corrects a docstring claiming the firmware steering law is
unreachable (Jupiter and unknown device types both land on it) and makes
that law read its power limits live, as the sibling law does, instead of
from a constructor snapshot the simulator mutates at runtime.

Metrics verified byte-identical over 33 scenarios x 3 seeds.
